+++ src/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { UserModule } from './user/user/user.module';
+import { UserModule } from './user/user.module';
 
 @Module({
   imports: [UserModule],
