+++ src/models.ts
@@ -1,17 +1,7 @@
-
-
 export interface Artist {
   id: string; // uuid v4
   name: string;
   grammy: boolean;
-}
-
-export interface Track {
-  id: string; // uuid v4
-  name: string;
-  artistId: string | null; // refers to Artist
-  albumId: string | null; // refers to Album
-  duration: number; // integer number
 }
 
 export interface Album {
